fix(review): address thermo-nuclear code-quality findings
Dead infrastructure (code-judo delete):
- Delete events.ts: runAfterCommit was an empty no-op, DomainEvent was an
  intermediate object that doubled the event-row description. Inline the
  insert values directly; the SEAM marker comment documents where packs
  append in-batch statements.
- Delete assertPostingAllowed: empty-body function with unused param.
  Removed opts.postingDate plumbing with it.

Half-wired schema:
- Per-line tax_amount/taxable_base are now computed and persisted at
  addLineItem time (was: computed in memory, discarded). Exported
  computeLineTax/computeLineTaxableBase from totals.ts; both the
  addLineItem write path and computeDocumentTotals now share them.

Authorization:
- finalize (freezes a fiscal record) was gated behind catalog:write (same
  as 'rename a product'). Added document:write permission (operator+);
  all document routes now use it. document:void stays admin+.

Boundary/contract (ADR-004):
- Agent tool duplicated the DOCUMENT_TYPES enum by hand — now reuses
  documentTypeSchema from contract (single source of truth).
- Hono GET / cast 'as DocumentType' — now uses zValidator with
  documentTypeSchema (validated at the boundary, not trusted).
- documentListItemSchema.total is now z.number().int() (money convention).

Correctness:
- UI totals used falsy coalescing (doc.subtotal || liveSubtotal) — a
  legit $0 finalized invoice fell through to live values. Now explicit
  isDraft ternary.

Type cleanup:
- FinalizeResult.folio is now number (not number | null) — the batch
  committed or we threw; the null was fictional.
- assertCanFinalize collapsed 3 status branches into 1 (draft-only).

Verified: typecheck 9/9, lint clean, tests 115 web + 20 registry.

+++ packages/contract/src/transaction.ts
@@ -61,7 +61,7 @@ export const documentListItemSchema = z.object({
   status: z.string(),
   entityName: z.string().nullable(),
   currencyCode: z.string(),
-  total: z.number(),
+  total: z.number().int(),
   folio: z.number().nullable(),
   createdAt: z.string(),
 });
